requirements.txt
openpyxl
numpy
selenium
keyboard
pytest
python-dotenv
pynput